03-production-deployment/agents/requirements.txt
strands-agents
boto3
uvicorn[standard]
fastapi
pydantic
requests
mcp
